setting, update and delete goals

+++ backend/controllers/goalController.js
@@ -1,77 +1,88 @@
-const asyncHandler = require("express-async-handler")
+const asyncHandler = require('express-async-handler')
 
-const Goal = require("../model/goalModel.js")
+const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
-//       GET
+// @desc    Get goals
+// @route   GET /api/goals
+// @access  Private
+const getGoals = asyncHandler(async (req, res) => {
+  const goals = await Goal.find({ user: req.user.id })
 
-// @ desc  Get Goals
-// @route  Get /api/goals
-// @acces  Private
-
-const getGoals = asyncHandler ( async (req,res) => {
-  const goals = await Goal.find()
   res.status(200).json(goals)
-
-  //console.log(req.body)
-  // res.status(200).json({message: "Get goals"})
 })
 
-// SET -POST
-
-// @ desc  set Goals
-// @route  POST /api/goals
-// @acces  Private
-const setGoal =  asyncHandler (async (req,res) => {
-  if(!req.body.text) {
+// @desc    Set goal
+// @route   POST /api/goals
+// @access  Private
+const setGoal = asyncHandler(async (req, res) => {
+  if (!req.body.text) {
     res.status(400)
-    throw new Error("Please add a text field")
+    throw new Error('Please add a text field')
   }
 
   const goal = await Goal.create({
-    text: req.body.text
+    text: req.body.text,
+    user: req.user.id,
   })
-
 
   res.status(200).json(goal)
 })
 
-// UPDATE
-
-// @ desc  Update Goal
-// @route  PUT /api/goals/:id
-// @acces  Private
-
-const updateGoal = asyncHandler ( async (req,res) => {
+// @desc    Update goal
+// @route   PUT /api/goals/:id
+// @access  Private
+const updateGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id)
 
-  if(!goal) {
+  if (!goal) {
     res.status(400)
     throw new Error('Goal not found')
+  }
+
+  // Check for user
+  if (!req.user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
-    new: true
+    new: true,
   })
+
   res.status(200).json(updatedGoal)
 })
 
-
-// DELETE 
-
-// @ desc  Delete Goals
-// @route  DELETE /api/goals/:id
-// @acces  Private
-
-const deleteGoal = asyncHandler( async (req,res) => {
+// @desc    Delete goal
+// @route   DELETE /api/goals/:id
+// @access  Private
+const deleteGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id)
 
-  if(!goal) {
+  if (!goal) {
     res.status(400)
     throw new Error('Goal not found')
   }
 
+  // Check for user
+  if (!req.user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
   await goal.deleteOne()
-  
 
   res.status(200).json({ id: req.params.id })
 })
@@ -80,5 +91,5 @@ module.exports = {
   getGoals,
   setGoal,
   updateGoal,
-  deleteGoal
+  deleteGoal,
 }
